# restful/docs/swagger.yaml
definitions:
...
  models.StoreProduct:
    properties:
      addTime:
        description: 添加时间
        type: integer
      brand:
        description: 品牌id
        type: integer
      browse:
        description: 浏览量
        type: integer
      cateId:
        description: 分类id
        type: string
      codePath:
        description: 产品二维码地址(用户小程序海报)
        type: string
      cost:
        description: 成本价
        type: number
      description:
        description: 产品描述
        type: string
      ficti:
        description: 虚拟销量
        type: integer
      giveIntegral:
        description: 获得积分
        type: number
      id:
        description: 商品id
        type: integer
      image:
        description: 商品图片
        type: string
      isBargain:
        description: 砍价状态 0未开启 1开启
        type: integer
      isBenefit:
        description: 是否优惠
        type: integer
      isBest:
        description: 是否精品
        type: integer
      isDel:
        description: 是否删除
        type: integer
      isHot:
        description: 是否热卖
        type: integer
      isNew:
        description: 是否新品
        type: integer
      isPostage:
        description: 是否包邮
        type: integer
      isSeckill:
        description: 秒杀状态 0 未开启 1已开启
        type: integer
      isShow:
        description: 状态（0：未上架，1：上架）
        type: integer
      keyword:
        description: 关键字
        type: string
      merId:
        description: 商户Id(0为总后台管理员创建,不为0的时候是商户后台创建)
        type: integer
      merUse:
        description: 商户是否代理 0不可代理1可代理
        type: integer
      otPrice:
        description: 市场价
        type: number
      postage:
        description: 邮费
        type: number
      price:
        description: 商品价格
        type: number
      sales:
        description: 销量
        type: integer
      sliderImage:
        description: 轮播图
        type: string
      sort:
        description: 排序
        type: integer
      soureLink:
        description: 淘宝、天猫、1688商品保存标识，避免商品重复入库
        type: string
      stock:
        description: 库存
        type: integer
      storeCode:
        description: 产品编码
        type: string
      storeInfo:
        description: 商品简介
        type: string
      storeName:
        description: 商品名称
        type: string
      unitName:
        description: 单位名
        type: string
      vipPrice:
        description: 会员价格
        type: number
    type: object
...
  models.StoreSeckill:
    properties:
      addTime:
        description: 添加时间
        type: string
      cost:
        description: 成本
        type: number
      description:
        description: 内容
        type: string
      giveIntegral:
        description: 返多少积分
        type: number
      id:
        description: 商品秒杀产品表id
        type: integer
      image:
        description: 推荐图
        type: string
      images:
        description: 轮播图
        type: string
      info:
        description: 简介
        type: string
      isDel:
        description: 删除 0未删除1已删除
        type: integer
      isHot:
        description: 热门推荐
        type: integer
      isPostage:
        description: 是否包邮
        type: integer
      isShow:
        description: 显示
        type: integer
      num:
        description: 最多秒杀几个
        type: integer
      otPrice:
        description: 原价
        type: number
      postage:
        description: 邮费
        type: number
      price:
        description: 价格
        type: number
      productId:
        description: 商品id
        type: integer
      sales:
        description: 销量
        type: integer
      sort:
        description: 排序
        type: integer
      startTime:
        description: 开始时间
        type: string
      status:
        description: 产品状态
        type: integer
      stock:
        description: 库存
        type: integer
      stopTime:
        description: 结束时间
        type: string
      title:
        description: 活动标题
        type: string
      unitName:
        description: 单位名
        type: string
    type: object
...
  models.StoreServiceLog:
    properties:
      addTime:
        description: 发送时间
        type: integer
      id:
        description: 客服用户对话记录表ID
        type: integer
      merId:
        description: 商户id
        type: integer
      msn:
        description: 消息内容
        type: string
      remind:
        description: 是否提醒过
        type: integer
      toUid:
        description: 接收人uid
        type: integer
      type:
        description: 是否已读（0：否；1：是；）
        type: integer
      uid:
        description: 发送人uid
        type: integer
    type: object
...
info:
  contact: {}
  license: {}
paths:
...
  /storeProducts:
    get:
      parameters:
      - description: query, ?query=age:>:21,name:like:%jason%
        in: query
        name: query
        type: string
      - description: order, ?order=age:desc,created_at:asc
        in: query
        name: order
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: pageSize
        in: query
        name: pageSize
        type: integer
      responses:
        "200":
          description: storeProduct array
          schema:
            items:
              $ref: '#/definitions/models.StoreProduct'
            type: array
      summary: List
      tags:
      - StoreProduct
    post:
      parameters:
      - description: StoreProduct
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.StoreProduct'
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Create
      tags:
      - StoreProduct
  /storeProducts/{storeProductId}:
    delete:
      parameters:
      - description: storeProductId
        in: path
        name: storeProductId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Delete
      tags:
      - StoreProduct
    get:
      parameters:
      - description: storeProductId
        in: path
        name: storeProductId
        required: true
        type: string
      responses:
        "200":
          description: storeProduct object
          schema:
            $ref: '#/definitions/models.StoreProduct'
      summary: Get
      tags:
      - StoreProduct
    patch:
      parameters:
      - description: storeProduct
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.StoreProduct'
      - description: storeProductId
        in: path
        name: storeProductId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Patch
      tags:
      - StoreProduct
    put:
      parameters:
      - description: storeProduct
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.StoreProduct'
      - description: storeProductId
        in: path
        name: storeProductId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Put
      tags:
      - StoreProduct
  /storeSeckills:
    get:
      parameters:
      - description: query, ?query=age:>:21,name:like:%jason%
        in: query
        name: query
        type: string
      - description: order, ?order=age:desc,created_at:asc
        in: query
        name: order
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: pageSize
        in: query
        name: pageSize
        type: integer
      responses:
        "200":
          description: storeSeckill array
          schema:
            items:
              $ref: '#/definitions/models.StoreSeckill'
            type: array
      summary: List
      tags:
      - StoreSeckill
    post:
      parameters:
      - description: StoreSeckill
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.StoreSeckill'
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Create
      tags:
      - StoreSeckill
  /storeSeckills/{storeSeckillId}:
    delete:
      parameters:
      - description: storeSeckillId
        in: path
        name: storeSeckillId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Delete
      tags:
      - StoreSeckill
    get:
      parameters:
      - description: storeSeckillId
        in: path
        name: storeSeckillId
        required: true
        type: string
      responses:
        "200":
          description: storeSeckill object
          schema:
            $ref: '#/definitions/models.StoreSeckill'
      summary: Get
      tags:
      - StoreSeckill
    patch:
      parameters:
      - description: storeSeckill
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.StoreSeckill'
      - description: storeSeckillId
        in: path
        name: storeSeckillId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Patch
      tags:
      - StoreSeckill
    put:
      parameters:
      - description: storeSeckill
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.StoreSeckill'
      - description: storeSeckillId
        in: path
        name: storeSeckillId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Put
      tags:
      - StoreSeckill
  /storeServiceLogs:
    get:
      parameters:
      - description: query, ?query=age:>:21,name:like:%jason%
        in: query
        name: query
        type: string
      - description: order, ?order=age:desc,created_at:asc
        in: query
        name: order
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: pageSize
        in: query
        name: pageSize
        type: integer
      responses:
        "200":
          description: storeServiceLog array
          schema:
            items:
              $ref: '#/definitions/models.StoreServiceLog'
            type: array
      summary: List
      tags:
      - StoreServiceLog
    post:
      parameters:
      - description: StoreServiceLog
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.StoreServiceLog'
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Create
      tags:
      - StoreServiceLog
  /storeServiceLogs/{storeServiceLogId}:
    delete:
      parameters:
      - description: storeServiceLogId
        in: path
        name: storeServiceLogId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Delete
      tags:
      - StoreServiceLog
    get:
      parameters:
      - description: storeServiceLogId
        in: path
        name: storeServiceLogId
        required: true
        type: string
      responses:
        "200":
          description: storeServiceLog object
          schema:
            $ref: '#/definitions/models.StoreServiceLog'
      summary: Get
      tags:
      - StoreServiceLog
    patch:
      parameters:
      - description: storeServiceLog
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.StoreServiceLog'
      - description: storeServiceLogId
        in: path
        name: storeServiceLogId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Patch
      tags:
      - StoreServiceLog
    put:
      parameters:
      - description: storeServiceLog
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.StoreServiceLog'
      - description: storeServiceLogId
        in: path
        name: storeServiceLogId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Put
      tags:
      - StoreServiceLog
...
swagger: "2.0"